Fixed 'Binding element 'children' implicitly has an any type' in ThemeSwitcher.tsx and Providers.tsx
